# swagger.yaml
openapi: 3.0.3
info:
  title: Nexus Project Management API
  description: |
    RESTful API for the Nexus Project Management System.
    
    ## Authentication
    This API uses JWT (JSON Web Token) for authentication. After registering or logging in, 
    you will receive a JWT token that must be included in the `Authorization` header for all 
    protected endpoints.
    
    ## Authorization
    The API implements role-based access control (RBAC) with two roles:
    - **ADMIN**: Full access to all resources
    - **USER**: Limited access to assigned resources only
    
    ## Features
    - User authentication and management
    - Project management with team assignments
    - Task management with status tracking
    - Role-based authorization
    - Pagination and search capabilities
  version: 1.0.0
  contact:
    name: API Support
    email: support@nexus-pm.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.nexus-pm.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from login or register endpoints.
        
        Example: `Authorization: Bearer <your_jwt_token>`

  # ============================================
  # REUSABLE SCHEMAS
  # ============================================
  schemas:
...
    UserRole:
      type: string
      enum:
        - ADMIN
        - USER
      description: User role for authorization
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          description: Unique user identifier
        name:
          type: string
          example: John Doe
          description: User's full name
        email:
          type: string
          format: email
          example: john.doe@example.com
          description: User's email address
        role:
          $ref: '#/components/schemas/UserRole'
        createdAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
          description: Last update timestamp

    # ========== PROJECT SCHEMAS ==========
...
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: Jane Doe
          description: Updated user name
        role:
          $ref: '#/components/schemas/UserRole'
          description: Updated user role (Admin only)
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
          description: Array of users
        total:
          type: integer
          example: 42
          description: Total number of users matching the query
        page:
          type: integer
          example: 1
          description: Current page number
        limit:
          type: integer
          example: 10
          description: Number of items per page
        totalPages:
          type: integer
          example: 5
          description: Total number of pages

    # ========== ERROR SCHEMAS ==========
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
          description: HTTP status code
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          example: Validation failed
          description: Error message or array of messages
        error:
          type: string
          example: Bad Request
          description: Error type
    UnauthorizedError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 401
        message:
          type: string
          example: Unauthorized
        error:
          type: string
          example: Unauthorized
    ForbiddenError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 403
        message:
          type: string
          example: Forbidden resource
        error:
          type: string
          example: Forbidden
    NotFoundError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: Resource not found
        error:
          type: string
          example: Not Found

# ============================================
# API ENDPOINTS
# ============================================
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Get all users with pagination and search
      description: |
        Retrieve all users with pagination and optional search.
        
        **Authentication Required:** Yes (JWT)
        
        **Authorization:** ADMIN only
        
        - Only admins can list all users
        - Passwords are excluded from the response
        - Supports pagination and search by name or email
      operationId: getAllUsers
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (1-indexed)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
          example: 10
        - name: search
          in: query
          description: Search query for name or email (case-insensitive)
          required: false
          schema:
            type: string
          example: john
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
              examples:
                paginatedUsers:
                  summary: Paginated user list
                  value:
                    users:
                      - id: 550e8400-e29b-41d4-a716-446655440000
                        name: John Doe
                        email: john.doe@example.com
                        role: USER
                        createdAt: "2024-01-15T10:30:00Z"
                        updatedAt: "2024-01-15T10:30:00Z"
                      - id: 660e8400-e29b-41d4-a716-446655440001
                        name: Jane Smith
                        email: jane.smith@example.com
                        role: ADMIN
                        createdAt: "2024-01-14T09:00:00Z"
                        updatedAt: "2024-01-14T09:00:00Z"
                    total: 42
                    page: 1
                    limit: 10
                    totalPages: 5
                searchResults:
                  summary: Search results
                  value:
                    users:
                      - id: 550e8400-e29b-41d4-a716-446655440000
                        name: John Doe
                        email: john.doe@example.com
                        role: USER
                        createdAt: "2024-01-15T10:30:00Z"
                        updatedAt: "2024-01-15T10:30:00Z"
                    total: 1
                    page: 1
                    limit: 10
                    totalPages: 1
        '400':
          description: Validation error (invalid query parameters)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden - User role not permitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get a user by ID
      description: |
        Retrieve a single user by their ID.
        
        **Authentication Required:** Yes (JWT)
        
        **Authorization:**
        - **ADMIN**: Can view any user
        - **USER**: Can only view their own profile
        
        - Password is excluded from the response
      operationId: getUserById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID (UUID)
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              examples:
                user:
                  summary: User details
                  value:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    name: John Doe
                    email: john.doe@example.com
                    role: USER
                    createdAt: "2024-01-15T10:30:00Z"
                    updatedAt: "2024-01-15T10:30:00Z"
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden - User cannot view this profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
              examples:
                notOwnProfile:
                  summary: User trying to view another user's profile
                  value:
                    statusCode: 403
                    message: You can only view your own profile
                    error: Forbidden
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Users
      summary: Update a user
      description: |
        Update an existing user.
        
        **Authentication Required:** Yes (JWT)
        
        **Authorization:**
        - **ADMIN**: Can update name and role for any user
        - **USER**: Can only update their own name (cannot change role)
        
        - All fields are optional (partial update)
        - Password cannot be updated through this endpoint
      operationId: updateUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID (UUID)
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
            examples:
              updateName:
                summary: Update user name (any user)
                value:
                  name: Jane Doe
              updateRole:
                summary: Update user role (admin only)
                value:
                  role: ADMIN
              updateBoth:
                summary: Update name and role (admin only)
                value:
                  name: Jane Doe
                  role: ADMIN
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              examples:
                updatedUser:
                  summary: Updated user
                  value:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    name: Jane Doe
                    email: john.doe@example.com
                    role: ADMIN
                    createdAt: "2024-01-15T10:30:00Z"
                    updatedAt: "2024-01-20T15:45:00Z"
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden - User cannot update this profile or role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
              examples:
                notOwnProfile:
                  summary: User trying to update another user
                  value:
                    statusCode: 403
                    message: You can only update your own profile
                    error: Forbidden
                userChangingRole:
                  summary: Regular user trying to change role
                  value:
                    statusCode: 403
                    message: Only admins can change user roles
                    error: Forbidden
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'